requirements.txt
Django==4.2.1
django-autoslug==1.9.9
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6